LPS-206234 Add notification info base to reuse code
